# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "JSON with login and password for registration",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User has been successfully registered",
            "content": {
              "application/json": {
                "example": {
                  "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
                  "refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjNkZmI4ZjA4LWQ1Y2EtNDc3Yy05M2I2LTE4N2U3OTZiOTI5NSIsImVtYWlsIjoidG90b3RvQGdtYWlsLmNvbSIsInN1YnNjcmlwdGlvbiI6ImZyZWUiLCJzdHJpcGVDdXN0b21lcklkIjpudWxsLCJyb2xlIjoiVVNFUiIsImlhdCI6MTc0NTU1ODg1OSwiZXhwIjoxNzQ2MTYzNjU5fQ.vBfGp9eFCBpfDqY3TkiDA8Vc4MF2OvI9bDk_456d91Q"
                }
              }
            }
          },
          "400": {
            "description": "User is already registered or incorrect property",
            "content": {
              "application/json": {
                "examples": {
                  "USER_ALREADY_REGISTERED": {
                    "summary": "This user is already registered",
                    "value": {
                      "message": "This user is already registered",
                      "error": "Bad Request",
                      "statusCode": 400
                    }
                  },
                  "PROPERTY_SHOULD_NOT_EXIST": {
                    "summary": "property should not exist",
                    "value": {
                      "message": "Property role should not exist",
                      "error": "Bad Request",
                      "statusCode": 400
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "summary": "Send your password and email as login for registration in the body of the request",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "JSON with login and password for Login",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in",
            "content": {
              "application/json": {
                "example": {
                  "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
                  "refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjNkZmI4ZjA4LWQ1Y2EtNDc3Yy05M2I2LTE4N2U3OTZiOTI5NSIsImVtYWlsIjoidG90b3RvQGdtYWlsLmNvbSIsInN1YnNjcmlwdGlvbiI6ImZyZWUiLCJzdHJpcGVDdXN0b21lcklkIjpudWxsLCJyb2xlIjoiVVNFUiIsImlhdCI6MTc0NTU1ODg1OSwiZXhwIjoxNzQ2MTYzNjU5fQ.vBfGp9eFCBpfDqY3TkiDA8Vc4MF2OvI9bDk_456d91Q"
                }
              }
            }
          },
          "400": {
            "description": "Unauthorized or incorrect property",
            "content": {
              "application/json": {
                "examples": {
                  "USER_NOT_FOUND_ERROR": {
                    "summary": "No user with this email found",
                    "value": {
                      "message": "No user with this email found",
                      "error": "Unauthorized",
                      "statusCode": 401
                    }
                  },
                  "PROPERTY_SHOULD_NOT_EXIST": {
                    "summary": "property should not exist",
                    "value": {
                      "message": "Property role should not exist",
                      "error": "Bad Request",
                      "statusCode": 400
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid password",
            "content": {
              "application/json": {
                "examples": {
                  "WRONG_PASSWORD_ERROR": {
                    "summary": "Invalid password",
                    "value": {
                      "message": "Invalid password",
                      "error": "Unauthorized",
                      "statusCode": 401
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "summary": "send your password and email as login in the request body",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Getting user profile",
            "content": {
              "application/json": {
                "example": {
                  "id": "3dfb8f08-d5ca-472c-93b6-187e796b9225",
                  "email": "test@gmail.com",
                  "subscription": "free",
                  "stripeCustomerId": null,
                  "currentTotalFolderSize": 54857600
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "message": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send your access_token as a BEARER token to headers authorization to get user profile",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Get new access and refresh token",
            "content": {
              "application/json": {
                "example": {
                  "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
                  "refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjNkZmI4ZjA4LWQ1Y2EtNDc3Yy05M2I2LTE4N2U3OTZiOTI5NSIsImVtYWlsIjoidG90b3RvQGdtYWlsLmNvbSIsInN1YnNjcmlwdGlvbiI6ImZyZWUiLCJzdHJpcGVDdXN0b21lcklkIjpudWxsLCJyb2xlIjoiVVNFUiIsImlhdCI6MTc0NTU1ODg1OSwiZXhwIjoxNzQ2MTYzNjU5fQ.vBfGp9eFCBpfDqY3TkiDA8Vc4MF2OvI9bDk_456d91Q"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "message": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "refresh_token": []
          }
        ],
        "summary": "Send old refresh_token as a BEARER token to headers authorization to get new access token and refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/signout": {
      "post": {
        "operationId": "AuthController_signOut",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Log out"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "message": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send the access_token as a BEARER token for authorization headers to log out of the account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/google/login": {
      "get": {
        "operationId": "AuthController_googleLogin",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "summary": "Endpoint for registration and login using google account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/google/callback": {
      "get": {
        "operationId": "AuthController_googleCallback",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "summary": "Endpoint to which redirection will occur after Google confirms the login. It will receive all the necessary data, create a user and redirect to the client page with the addition of searchParameters",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Fl studio API",
    "description": "Documentation for Fl studio API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access_token": {
        "scheme": "Bearer",
        "bearerFormat": "Bearer",
        "description": "Please enter token in following format: Bearer <JWT>",
        "name": "Authorization",
        "type": "http",
        "in": "Header"
      },
      "refresh_token": {
        "scheme": "Bearer",
        "bearerFormat": "Bearer",
        "description": "Refresh token: Bearer <JWT>",
        "name": "Authorization",
        "type": "http",
        "in": "Header"
      }
    },
    "schemas": {
      "AuthDto": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "description": "User email",
            "example": "test@gmail.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "password123"
          }
        },
        "required": [
          "login",
          "password"
        ]
      },
...
    }
  }
}